refactor: extract internal/versions package from server
Move version fingerprint, ingest, swap, and lifecycle logic from
internal/server into internal/versions. The server package retains
thin wrappers and HTTP handlers that delegate to the new package.

RunVersionSwap now takes onWriteBack and notifyDeluge callbacks
instead of direct *WriteBackBatcher dependency, decoupling the
versions package from server internals.
